updated searchindex and removed old search example
