--- Examples/Search/config.wyam
@@ -1,28 +1,0 @@
-Pipelines.Add("Posts",
-	ReadFiles(@"posts\*.md"),  // Read all markdown files in the "posts" directory
-	FrontMatter(Yaml()),  // Load any frontmatter and parse it as YAML markup
-	Markdown(),  // Render the markdown content
-	Meta("Post", @doc.Content),  // Move the markdown content to metadata
-	ReadFiles(@"posts\Post.cshtml"),  // Load the Razor post page template
-	Razor(),  // Compile and render the page template
-	Meta("PostFile", string.Format(@"posts\{0}.html", ((string)@doc["Title"]).ToLower().Replace(' ', '-'))),  // Use the post title as the filename and save it in metadata so we can use it in links later
-	WriteFiles((string)@doc["PostFile"]),  // Write the post file 
-	Meta("SearchIndexItem", new SearchIndexItem(url: (string)@doc["PostFile"],
-												title: (string)@doc["Title"],
-												content: @doc.Content)
-		)
-);
-
-Pipelines.Add("Home",
-	ReadFiles(@"*.cshtml").FromTopDirectoryOnly().Where(x => Path.GetFileName(x)[0] != '_'),
-	Razor(),
-	WriteFiles(".html")
-);
-
-Pipelines.Add("Copy",
-	CopyFiles(@"*.js")
-);
-
-Pipelines.Add("Search",
-	SearchIndex("js/searchindex.js", "stopwords.txt")
-);
