--- Examples/Search/input/_Layout.cshtml
@@ -1,5 +1,0 @@
-THIS IS MY PAGE
-<hr />
-@RenderBody()
-<hr />
-FOOTER
