--- Examples/Search/input/home.cshtml
@@ -1,10 +1,0 @@
-<h1>Home</h1>
-<hr/>
-<p>Hi, this paragraph has a little bit about me.</p>
-<h2>Posts</h2>
-@foreach(var doc in Documents["Posts"].OrderBy(x => x["Title"]))
-{
-	<ul>
-		<li><a href="@doc["PostFile"]">@doc["Title"]</a> - @doc["Published"]</li>
-	</ul>
-}
