--- Examples/Search/input/posts/a.md
@@ -1,7 +1,0 @@
-Published: 2015-04-25
-Title: This Is My First Post
----
-Welcome to my blog. This is my very **first post**. It is called dependency-injection :-)
-I like C#
-
-I "knew" it. It's nice
