--- Examples/Search/input/posts/b.md
@@ -1,12 +1,0 @@
-Published: 2015-04-28
-Title: Another Post
----
-And here's another post.
-
-# Heading 1
-
-This post has some great content!
-
-## Heading 2
-
-And it's got subheadings!
