--- Examples/Search/input/posts/post.cshtml
@@ -1,4 +1,0 @@
-<h1>@Metadata["Title"]</h1>
-<h2>@Metadata["Published"]</h2>
-<hr />
-@Html.Raw(Metadata["Post"])
