--- Examples/Search/input/search.cshtml
@@ -1,38 +1,0 @@
-<h1>Search</h1>
-<hr/>
-<input type="search" id="search" autofocus="autofocus" placeholder="Search..."/>
-<div id="searchResultPanel"></div>
-<script src="/js/lunr.min.js"></script>
-<script src="/js/searchindex.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/1.11.3/jquery.js"></script>
-
-<script>
-function getQueryParam(name){
-    var results = new RegExp('[\?&]' + name + '=([^&#]*)').exec(window.location.href);
-    if (results==null){
-        return null;
-    }
-    else{
-        return results[1] || 0;
-    }
-}
-
-$(document).ready(function(){
-	$("#search").on('input propertychange paste', function() {
-		var query = $("#search").val();
-		if( query.length < 3 )
-			return;
-		
-		var result = searchModule.search(query);
-		console.log(result);
-	});
-	
-	var q = getQueryParam("q");
-    
-    if( q )
-    {
-        q = decodeURIComponent(q);
-        $("#search").val(q);
-	}
-});
-</script>
