--- Examples/Search/input/stopwords.txt
@@ -1,11 +1,0 @@
-it
-is
-you
-has
-had
-have
-got
-this
-my
-to
-and
