+++ Examples/SearchIndex/input/_Layout.cshtml
@@ -1,7 +1,7 @@
 <html>
 	<head>
 		<title>Generated Blog - @Model["Title"]</title>
-		<script src="/assets/js/jquery.js"></script>
+		<script src="/assets/js/jquery.min.js"></script>
 	</head>
 	<body>
 		<h1>My Generated Blog!</h1>
