+++ Wyam.Modules.SearchIndex/MetadataKeys.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Wyam.Modules.SearchIndex
+{
+    public static class MetadataKeys
+    {
+        public const string SearchIndexItem = "SearchIndexItem";
+    }
+}
